+++ src/com/landawn/abacus/util/OptionalChar.java
@@ -209,15 +209,19 @@ public final class OptionalChar implements Comparable<OptionalChar> {
     }
 
     /**
+     * Same as {@code orElseZero}.
      * 
-     * @return
-     * @deprecated replaced with orElseZero.
+     * @return.
      */
-    @Deprecated
     public char orZero() {
         return isPresent() ? value : 0;
     }
 
+    /**
+     * Same as {@code orZero}.
+     * 
+     * @return.
+     */
     public char orElseZero() {
         return isPresent() ? value : 0;
     }