+++ src/com/landawn/abacus/util/OptionalInt.java
@@ -213,15 +213,19 @@ public final class OptionalInt implements Comparable<OptionalInt> {
     }
 
     /**
+     * Same as {@code orElseZero}.
      * 
-     * @return
-     * @deprecated replaced with orElseZero.
+     * @return.
      */
-    @Deprecated
     public int orZero() {
         return isPresent() ? value : 0;
     }
 
+    /**
+     * Same as {@code orZero}.
+     * 
+     * @return.
+     */
     public int orElseZero() {
         return isPresent() ? value : 0;
     }