+++ src/com/landawn/abacus/util/OptionalLong.java
@@ -213,15 +213,19 @@ public final class OptionalLong implements Comparable<OptionalLong> {
     }
 
     /**
+     * Same as {@code orElseZero}.
      * 
-     * @return
-     * @deprecated replaced with orElseZero.
+     * @return.
      */
-    @Deprecated
     public long orZero() {
         return isPresent() ? value : 0;
     }
 
+    /**
+     * Same as {@code orZero}.
+     * 
+     * @return.
+     */
     public long orElseZero() {
         return isPresent() ? value : 0;
     }